+++ common/src/main/java/io/github/axolotlclient/modules/hud/gui/hud/simple/PlayerCountHud.java
@@ -22,6 +22,7 @@
 
 package io.github.axolotlclient.modules.hud.gui.hud.simple;
 
+import io.github.axolotlclient.AxolotlClientCommon;
 import io.github.axolotlclient.bridge.util.AxoI18n;
 import io.github.axolotlclient.bridge.util.AxoIdentifier;
 import io.github.axolotlclient.modules.hud.gui.entry.SimpleTextHudEntry;
@@ -35,7 +36,7 @@ import io.github.axolotlclient.modules.hud.gui.entry.SimpleTextHudEntry;
 
 public class PlayerCountHud extends SimpleTextHudEntry {
 
-	public static final AxoIdentifier ID = AxoIdentifier.of("axolotlclient", "playercounthud");
+	public static final AxoIdentifier ID = AxoIdentifier.of(AxolotlClientCommon.MODID, "playercounthud");
 
 	public PlayerCountHud() {
 		super(true);